add routes and router
