+++ src/pages/login/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function LoginPage() {
+  return <div>LoginPage</div>;
+}
+
+export default LoginPage;
