+++ src/pages/signup/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SignUpPage() {
+  return <div>SignUpPage</div>;
+}
+
+export default SignUpPage;
