+++ src/services/index.js
@@ -1,0 +1,2 @@
+import AuthService from './AuthService';
+export { AuthService };
